requirements.txt

torch>=2.2
torchvision>=0.17
numpy
opencv-python
scikit-learn
pyyaml
tqdm
pydicom
einops
